Move SystemStats to the Statistics module
